requirements.txt
websocket-client>=0.44.0
six>=1.10.0
